major render bug fixes
Fixed some issues where not all tokens were being evaluated, depending
on which token was evaluated first in the tier. Also changed
'hasOwnProperty' to 'in' so that the object_rsearch function can
traverse through the prototype. This is especially important for using
Bowtie in conjunction with a data management tool like Fishtank.js
(Although I will update it to serve up non-prototypal values eventually)