Bowtie v2.0.0
Changed the entire mechanism by which the templates are both compiled
and parsed and (for the most part) abandoned the RegExp style of
rendering. The result is a performance level that is nearly 1/50 of what
it used to be and benchmarks for in-browser compiling (once) and
rendering (in a loop) float a little bit below the typical mustache.js
level. I will work to improve performance further using pre-compilation,
but in the meantime, the new features should outweigh the tiny
difference in speed between bowtie.js and handlebars.js